requirements.txt
# Core dependencies
crewai>=0.80.0
pydantic>=2.0.0
pydantic-ai>=0.0.15
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.0.0

# Data processing and utilities
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Async support
asyncio-mqtt>=0.16.0
aiofiles>=23.0.0

# Logging and monitoring
structlog>=23.0.0
rich>=13.0.0

# Optional integrations
redis>=4.5.0
sqlalchemy>=2.0.0
networkx>=3.0

# Development tools (optional)
jupyter>=1.0.0
ipython>=8.0.0